pyproject.toml
[project]
name = "hurricane"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cdsapi>=0.7.7",
    "dask>=2026.3.0",
    "herbie-data>=2026.3.0",
    "netcdf4>=1.7.4",
    "neurovolume>=0.1.0a19",
    "numpy>=2.4.4",
    "scipy>=1.17.1",
    "xarray>=2026.4.0",
]
